Refactoring to prepare state buffer.
