+++ shared/src/lib.rs
@@ -1,7 +1,11 @@
-#![feature(core)]
+#![feature(plugin)]
+#![plugin(phf_macros)]
+
+extern crate phf;
 
 pub mod asset;
 pub mod resource;
 pub mod style;
 pub mod deps;
 pub mod markup;
+pub mod properties;
